LayoutTests/media/modern-media-controls/icon-button/icon-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>IconButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const layoutDelegate = {};

const iconButton = new IconButton({ layoutDelegate });

shouldBeEqualToString("iconButton.element.localName", "button");
shouldBeEqualToString("iconButton.element.className", "icon");

debug("");
debug("There should be no iconName by default");
shouldBeEqualToString("iconButton.iconName", "");

debug("");
debug("Setting an iconName with invalid layout traits should throw");
shouldThrow(function() { iconButton.iconName = Icons.Pause });
shouldBeEqualToString("iconButton.iconName", "");

debug("");
debug("Setting macOS inline layout traits and setting play icon");
layoutDelegate.layoutTraits = LayoutTraits.macOS;
iconButton.iconName = Icons.Pause;

let numberOfFrames = 0;

scheduler.frameDidFire = function()
{
    numberOfFrames++;

    if (numberOfFrames === 1) {
        shouldBeTrue('iconButton.element.style.webkitMaskImage.includes("macOS/pause")');
        shouldBeEqualToString("iconButton.element.style.width", "22px");
        shouldBeEqualToString("iconButton.element.style.height", "24px");
        shouldBeEqualToString("iconButton.element.style.webkitMaskSize", "22px 24px");

        debug("");
        debug("Setting start icon");
        iconButton.iconName = Icons.Start;
    } else if (numberOfFrames === 2) {
        shouldBeTrue('iconButton.element.style.webkitMaskImage.includes("macOS/start")');
        shouldBeEqualToString("iconButton.element.style.width", "70px");
        shouldBeEqualToString("iconButton.element.style.height", "70px");
        shouldBeEqualToString("iconButton.element.style.webkitMaskSize", "70px 70px");

        finishJSTest();
    } 
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
